pyproject.toml
[project]
name = "template-mcp"
version = "0.1.0"
description = "MCP server for structured markdown templates"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=0.9.0",
    "pyyaml>=6.0"
]

[project.scripts]
template-mcp = "template_mcp.server:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
